+++ store.go
@@ -1,6 +1,6 @@
 package main
 
-import tg "github.com/semog/go-bot-api/v4"
+import tg "github.com/semog/go-bot-api/v5"
 
 // Store is an interface for the persistent storage
 // should allow easier swapping of databases
@@ -10,22 +10,22 @@ type Store interface {
 	AddMsgToPoll(pollID int, messageID int, chatID int64) error
 	AddInlineMsgToPoll(pollID int, inlineMessageID string) error
 	RemoveInlineMsg(inlineMessageID string) error
-	GetUser(userID int) (*tg.User, error)
+	GetUser(userID int64) (*tg.User, error)
 	GetPoll(pollID int) (*poll, error)
-	GetUserPoll(pollID int, userID int) (*poll, error)
-	GetPollsByUser(userID int) ([]*poll, error)
+	GetUserPoll(pollID int, userID int64) (*poll, error)
+	GetPollsByUser(userID int64) ([]*poll, error)
 	GetPollID(messageID int) (int, error)
-	GetPollNewer(pollID int, userID int) (*poll, error)
-	GetPollOlder(pollID int, userID int) (*poll, error)
+	GetPollNewer(pollID int, userID int64) (*poll, error)
+	GetPollOlder(pollID int, userID int64) (*poll, error)
 	GetAllPollMsg(pollID int) ([]pollident, error)
 	GetAllPollInlineMsg(pollID int) ([]pollident, error)
-	GetState(userID int) (state int, pollID int, err error)
-	SaveState(userID int, pollID int, state int) error
+	GetState(userID int64) (state int, pollID int, err error)
+	SaveState(userID int64, pollID int, state int) error
 	SaveUser(*tg.User) error
 	SavePoll(*poll) (int, error)
 	SaveOptions([]option) error
 	DeleteOptions([]option) error
 	SaveAnswer(*poll, answer) (unvoted bool, err error)
-	ResetPoll(userID int, pollID int) error
-	DeletePoll(userID int, pollID int) error
+	ResetPoll(userID int64, pollID int) error
+	DeletePoll(userID int64, pollID int) error
 }
